Merge pull request

+++ app/Traits/Permissions.php
@@ -4,12 +4,18 @@ namespace App\Traits;
 
 use App\Models\Auth\Permission;
 use App\Models\Auth\Role;
+use App\Traits\SearchString;
+use App\Traits\Translations;
 use App\Utilities\Reports;
 use App\Utilities\Widgets;
+use Illuminate\Routing\Route;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 
 trait Permissions
 {
+    use SearchString, Translations;
+
     public function getActionsMap()
     {
         return [
@@ -33,42 +39,59 @@ trait Permissions
 
     public function attachPermissionsToAdminRoles($permissions)
     {
-        $this->attachPermissionsToAllRoles($permissions, 'read-admin-panel');
+        $this->applyPermissionsToRoles($this->getDefaultAdminRoles(), 'attach', $permissions);
     }
 
     public function attachPermissionsToPortalRoles($permissions)
     {
-        $this->attachPermissionsToAllRoles($permissions, 'read-client-portal');
+        $this->applyPermissionsToRoles($this->getDefaultPortalRoles(), 'attach', $permissions);
     }
 
     public function attachPermissionsToAllRoles($permissions, $require = 'read-admin-panel')
     {
-        $this->getRoles($require)->each(function ($role) use ($permissions) {
-            foreach ($permissions as $id => $permission) {
-                if ($this->isActionList($permission)) {
-                    $this->attachPermissionsByAction($role, $id, $permission);
-
-                    continue;
-                }
-
-                $this->attachPermission($role, $permission);
-            }
-        });
+        $this->applyPermissionsToRoles($this->getRoles($require), 'attach', $permissions);
     }
 
     public function detachPermissionsByRoleNames($roles)
     {
         foreach ($roles as $role_name => $permissions) {
-            $role = Role::where('name', $role_name)->first();
-
-            if (empty($role)) {
-                continue;
-            }
-
             foreach ($permissions as $permission_name) {
-                $this->detachPermission($role, $permission_name);
+                $this->detachPermission($role_name, $permission_name);
             }
         }
+    }
+
+    public function detachPermissionsFromAdminRoles($permissions)
+    {
+        $this->applyPermissionsToRoles($this->getDefaultAdminRoles(), 'detach', $permissions);
+    }
+
+    public function detachPermissionsFromPortalRoles($permissions)
+    {
+        $this->applyPermissionsToRoles($this->getDefaultPortalRoles(), 'detach', $permissions);
+    }
+
+    public function detachPermissionsFromAllRoles($permissions, $require = 'read-admin-panel')
+    {
+        $this->applyPermissionsToRoles($this->getRoles($require), 'detach', $permissions, $require);
+    }
+
+    public function applyPermissionsToRoles($roles, $apply, $permissions)
+    {
+        $roles->each(function ($role) use ($apply, $permissions) {
+            $f1 = $apply . 'PermissionsByAction';
+            $f2 = $apply . 'Permission';
+
+            foreach ($permissions as $id => $permission) {
+                if ($this->isActionList($permission)) {
+                    $this->$f1($role, $id, $permission);
+
+                    continue;
+                }
+
+                $this->$f2($role, $permission);
+            }
+        });
     }
 
     public function updatePermissionNames($permissions)
@@ -96,7 +119,7 @@ trait Permissions
         }
     }
 
-    public function attachDefaultModulePermissions($module, $require = 'read-admin-panel')
+    public function attachDefaultModulePermissions($module, $require = null)
     {
         $this->attachModuleReportPermissions($module, $require);
 
@@ -105,7 +128,7 @@ trait Permissions
         $this->attachModuleSettingPermissions($module, $require);
     }
 
-    public function attachModuleReportPermissions($module, $require = 'read-admin-panel')
+    public function attachModuleReportPermissions($module, $require = null)
     {
         if (is_string($module)) {
             $module = module($module);
@@ -125,10 +148,12 @@ trait Permissions
             $permissions[] = $this->createModuleReportPermission($module, $class);
         }
 
-        $this->attachPermissionsToAllRoles($permissions, $require);
+        $require
+                ? $this->attachPermissionsToAllRoles($permissions, $require)
+                : $this->attachPermissionsToAdminRoles($permissions);
     }
 
-    public function attachModuleWidgetPermissions($module, $require = 'read-admin-panel')
+    public function attachModuleWidgetPermissions($module, $require = null)
     {
         if (is_string($module)) {
             $module = module($module);
@@ -148,10 +173,12 @@ trait Permissions
             $permissions[] = $this->createModuleWidgetPermission($module, $class);
         }
 
-        $this->attachPermissionsToAllRoles($permissions, $require);
+        $require
+                ? $this->attachPermissionsToAllRoles($permissions, $require)
+                : $this->attachPermissionsToAdminRoles($permissions);
     }
 
-    public function attachModuleSettingPermissions($module, $require = 'read-admin-panel')
+    public function attachModuleSettingPermissions($module, $require = null)
     {
         if (is_string($module)) {
             $module = module($module);
@@ -166,7 +193,9 @@ trait Permissions
         $permissions[] = $this->createModuleSettingPermission($module, 'read');
         $permissions[] = $this->createModuleSettingPermission($module, 'update');
 
-        $this->attachPermissionsToAllRoles($permissions, $require);
+        $require
+                ? $this->attachPermissionsToAllRoles($permissions, $require)
+                : $this->attachPermissionsToAdminRoles($permissions);
     }
 
     public function createModuleReportPermission($module, $class)
@@ -220,13 +249,37 @@ trait Permissions
 
     public function createRole($name, $display_name = null, $description = null)
     {
-        $display_name = $display_name ?? Str::title($name);
+        $alias = !empty($this->alias) ? $this->alias : $name;
+
+        if (empty($display_name)) {
+            $display_name = $this->findTranslation([
+                'auth.roles.' . Str::replace('-', '_', $name) . '.name',
+                $alias . '::permissions.roles.' . Str::replace('-', '_', $name) . '.name',
+                $alias . '::auth.roles.' . Str::replace('-', '_', $name) . '.name',
+            ]);
+
+            if (empty($display_name)) {
+                $display_name = Str::title(Str::replace('-', ' ', $name));
+            }
+        }
+
+        if (empty($description)) {
+            $description = $this->findTranslation([
+                'auth.roles.' . Str::replace('-', '_', $name) . '.description',
+                $alias . '::permissions.roles.' . Str::replace('-', '_', $name) . '.description',
+                $alias . '::auth.roles.' . Str::replace('-', '_', $name) . '.description',
+            ]);
+
+            if (empty($description)) {
+                $description = $display_name;
+            }
+        }
 
         return Role::firstOrCreate([
             'name' => $name,
         ], [
             'display_name' => $display_name,
-            'description' => $description ?? $display_name,
+            'description' => $description,
         ]);
     }
 
@@ -255,7 +308,7 @@ trait Permissions
         $role->attachPermission($permission);
     }
 
-    public function detachPermission($role, $permission)
+    public function detachPermission($role, $permission, $delete = true)
     {
         if (is_string($role)) {
             $role = Role::where('name', $role)->first();
@@ -273,7 +326,14 @@ trait Permissions
             return;
         }
 
-        $role->detachPermission($permission);
+        if ($role->hasPermission($permission->name)) {
+            $role->detachPermission($permission);
+        }
+
+        if ($delete === false) {
+            return;
+        }
+
         $permission->delete();
     }
 
@@ -289,6 +349,18 @@ trait Permissions
 
     public function attachPermissionsByAction($role, $page, $action_list)
     {
+        $this->applyPermissionsByAction('attach', $role, $page, $action_list);
+    }
+
+    public function detachPermissionsByAction($role, $page, $action_list)
+    {
+        $this->applyPermissionsByAction('detach', $role, $page, $action_list);
+    }
+
+    public function applyPermissionsByAction($apply, $role, $page, $action_list)
+    {
+        $function = $apply . 'Permission';
+
         $actions_map = collect($this->getActionsMap());
 
         $actions = explode(',', $action_list);
@@ -298,7 +370,7 @@ trait Permissions
 
             $name = $action . '-' . $page;
 
-            $this->attachPermission($role, $name);
+            $this->$function($role, $name);
         }
     }
 
@@ -322,5 +394,122 @@ trait Permissions
         return Role::all()->filter(function ($role) use ($require) {
             return $require ? $role->hasPermission($require) : true;
         });
+    }
+
+    public function getDefaultAdminRoles($custom = null)
+    {
+        $roles = Role::whereIn('name', $custom ?? ['admin', 'manager'])->get();
+
+        if ($roles->isNotEmpty()) {
+            return $roles;
+        }
+
+        return $this->getRoles('read-admin-panel');
+    }
+
+    public function getDefaultPortalRoles($custom = null)
+    {
+        $roles = Role::whereIn('name', $custom ?? ['customer'])->get();
+
+        if ($roles->isNotEmpty()) {
+            return $roles;
+        }
+
+        return $this->getRoles('read-client-portal');
+    }
+
+    /**
+     * Assign permissions middleware to default controller methods.
+     *
+     * @return void
+     */
+    public function assignPermissionsToController()
+    {
+        // No need to check for permission in console
+        if (app()->runningInConsole()) {
+            return;
+        }
+
+        $table = request()->isApi() ? request()->segment(2) : '';
+
+        // Find the proper controller for common API endpoints
+        if (in_array($table, ['contacts', 'documents'])) {
+            $controller = '';
+
+            // Look for type in search variable like api/contacts?search=type:customer
+            $type = $this->getSearchStringValue('type');
+
+            if (! empty($type)) {
+                $alias = config('type.' . Str::singular($table) . '.' . $type . '.alias');
+                $group = config('type.' . Str::singular($table) . '.' . $type . '.group');
+                $prefix = config('type.' . Str::singular($table) . '.' . $type . '.permission.prefix');
+
+                // if use module set module alias
+                if (! empty($alias)) {
+                    $controller .= $alias . '-';
+                }
+
+                // if controller in folder it must
+                if (! empty($group)) {
+                    $controller .= $group . '-';
+                }
+
+                $controller .= $prefix;
+            }
+        } else {
+            $route = app(Route::class);
+
+            // Get the controller array
+            $arr = array_reverse(explode('\\', explode('@', $route->getAction()['uses'])[0]));
+
+            $controller = '';
+
+            // Add module
+            if (isset($arr[3]) && isset($arr[4])) {
+                if (strtolower($arr[4]) == 'modules') {
+                    $controller .= Str::kebab($arr[3]) . '-';
+                } elseif (isset($arr[5]) && (strtolower($arr[5]) == 'modules')) {
+                    $controller .= Str::kebab($arr[4]) . '-';
+                }
+            }
+
+            // Add folder
+            if (! in_array(strtolower($arr[1]), ['api', 'controllers'])) {
+                $controller .= Str::kebab($arr[1]) . '-';
+            }
+
+            // Add file
+            $controller .= Str::kebab($arr[0]);
+
+            // Skip ACL
+            $skip = ['portal-dashboard'];
+            if (in_array($controller, $skip)) {
+                return;
+            }
+
+            // App\Http\Controllers\FooBar                  -->> foo-bar
+            // App\Http\Controllers\FooBar\Main             -->> foo-bar-main
+            // Modules\Blog\Http\Controllers\Posts          -->> blog-posts
+            // Modules\Blog\Http\Controllers\Portal\Posts   -->> blog-portal-posts
+        }
+
+        // Add CRUD permission check
+        $this->middleware('permission:create-' . $controller)->only('create', 'store', 'duplicate', 'import');
+        $this->middleware('permission:read-' . $controller)->only('index', 'show', 'edit', 'export');
+        $this->middleware('permission:update-' . $controller)->only('update', 'enable', 'disable');
+        $this->middleware('permission:delete-' . $controller)->only('destroy');
+    }
+
+    public function canAccessMenuItem($title, $permissions)
+    {
+        $permissions = Arr::wrap($permissions);
+
+        $item = new \stdClass();
+        $item->title = $title;
+        $item->permissions = $permissions;
+
+        event(new \App\Events\Menu\ItemAuthorizing($item));
+
+        return user()->canAny($item->permissions);
     }
 }
